+++ view/adminhtml/requirejs-config.js
@@ -1,6 +1,6 @@
 /**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
+ * Copyright © 2023 Studio Raz. All rights reserved.
+ * See LICENSE for license details.
  */
 
 var config = {
@@ -8,6 +8,9 @@ var config = {
         mixins: {
             'Magento_PageBuilder/js/modal/modal': {
                 'SR_PageBuilder/js/modal/modal-mixin': true
+            },
+            'Magento_PageBuilder/js/content-type/style-registry': {
+                'SR_PageBuilder/js/content-type/style-registry-mixin': true
             }
         }
     }