+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.1] - 2023-08-10
+### Added
+- Implemented PB Columns grid
+- Implemented fieldset with description
+
 ## [1.0.0] - 2023-03-02
 ### Added
 - Project init
@@ -13,5 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add configuration for the font weight to the Page Builder
 
 
-[Unreleased]: https://github.com/studioraz/magento2-page-builder/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/studioraz/magento2-page-builder/compare/1.3.1...HEAD
+[1.3.1]: https://github.com/studioraz/magento2-page-builder/compare/1.3.0..1.3.1
 [1.0.0]: https://github.com/studioraz/magento2-page-builder/releases/tag/1.0.0